+++ cmd/hidden_lake/applications/filesharer/_daemon/disable_hlf.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+# root mode
+systemctl disable hidden_lake_filesharer.service