+++ cmd/hidden_lake/applications/remoter/_daemon/disable_hlr.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+# root mode
+systemctl disable hidden_lake_remoter.service