+++ cmd/hidden_lake/applications/remoter/_daemon/journal_hlr.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+journalctl -eu hidden_lake_remoter.service