+++ cmd/hidden_lake/applications/remoter/_daemon/stop_hlr.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+# root mode
+systemctl stop hidden_lake_remoter.service